Create comments and fix edge case in content transformation
Regarding the noted edge case, there are times where folks have used number tags
to refer to trello cards. We need to map those back to trello URLs so that we can
map those to GitHub issues in the card-linking pass of the content.

Made some of the promise helpers generic and overall did some refactoring.